requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.3
google-generativeai==0.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
pillow==10.2.0
